+++ src/messages.ts
@@ -1,7 +1,7 @@
-import { Tip, TipType } from './tip';
+import { Command, Tip, TipType } from './tip';
 
 export const error = (title: string, str: string): Tip => {
-	return new Tip(title, str, TipType.Error);
+	return new Tip(title, str, TipType.Error, str, Command.NoOp, 'OK').canIgnore();
 };
 
 export const libString = (lib: string, ver?: string) => {
@@ -16,7 +16,7 @@ export const writeMinVersionError = (library: string, version: string, minVersio
 		TipType.Error, undefined,
 		`npm install ${library}@latest --save-exact`,
 		`Upgrade`,
-		`${library} successfully updated.`);
+		`${library} successfully updated.`).canIgnore();
 };
 export const writeMinVersionWarning = (library: string, version: string, minVersion: string, reason: string, url?: string): Tip => {
 	let r = reason ? ' ' + reason : '';
@@ -26,7 +26,7 @@ export const writeMinVersionWarning = (library: string, version: string, minVers
 		`${library} ${version} should be updated to at least ${minVersion}${reason ? ' ' + reason : ''}`,
 		`npm install ${library}@latest --save-exact`,
 		`Upgrade`,
-		`${library} successfully updated.`);
+		`${library} successfully updated.`).canIgnore();
 };
 
 export const writeConsistentVersionWarning = (lib1: string, ver1: string, lib2: string, ver2: string) => {
@@ -36,7 +36,7 @@ export const writeConsistentVersionWarning = (lib1: string, ver1: string, lib2: 
 		TipType.Error, undefined,
 		`npm install ${lib2}@${ver1} --save-exact`,
 		`Upgrade`,
-		`${lib2} successfully updated.`);
+		`${lib2} successfully updated.`).canIgnore();
 };
 
 export const writeConsistentVersionError = (lib1: string, ver1: string, lib2: string, ver2: string): Tip => {
@@ -46,5 +46,5 @@ export const writeConsistentVersionError = (lib1: string, ver1: string, lib2: st
 		TipType.Error, undefined,
 		`npm install ${lib2}@${ver1} --save-exact`,
 		`Upgrade`,
-		`${lib2} successfully updated.`);
+		`${lib2} successfully updated.`).canIgnore();
 };